+++ subcommand/envoy-entrypoint/task-monitor.go
@@ -20,6 +20,7 @@ import (
 var (
 	nonAppContainers = map[string]struct{}{
 		"consul-ecs-control-plane": {},
+		"consul-dataplane":         {},
 	}
 )
 
@@ -100,6 +101,12 @@ func allAppContainersStopped(taskMeta awsutil.ECSTaskMeta) bool {
 }
 
 func isApplication(container awsutil.ECSTaskMetaContainer) bool {
+	// ECS might auto-inject some containers that aren't part of
+	// the task definition. We treat them as non app containers.
+	if !container.IsNormalType() {
+		return false
+	}
+
 	_, ok := nonAppContainers[container.Name]
 	return !ok
 }